Overview – Constants, Types, Constructors and Methods

Constants, Types and Constructors

ChargeTransport.BoundaryModelType

ChargeTransport.BulkRecombination

ChargeTransport.CalculationType

ChargeTransport.Data

ChargeTransport.Data

ChargeTransport.DiffusionEnhanced

ChargeTransport.ExcessChemicalPotential

ChargeTransport.ExcessChemicalPotentialGraded

ChargeTransport.FluxApproximationType

ChargeTransport.GeneralizedSG

ChargeTransport.GenerationBeerLambert

ChargeTransport.GenerationModelType

ChargeTransport.GenerationNone

ChargeTransport.GenerationUniform

ChargeTransport.InEquilibrium

ChargeTransport.InterfaceModelNone

ChargeTransport.InterfaceModelSurfaceReco

ChargeTransport.InterfaceModelType

ChargeTransport.IonicChargeCarriers

ChargeTransport.ModelType

ChargeTransport.OhmicContact

ChargeTransport.OutOfEquilibrium

ChargeTransport.OuterBoundaryModelType

ChargeTransport.Params

ChargeTransport.Params

ChargeTransport.ParamsNodal

ChargeTransport.ParamsNodal

ChargeTransport.QType

ChargeTransport.SRHWithTrapsType

ChargeTransport.SRHWithoutTrapsType

ChargeTransport.ScharfetterGummel

ChargeTransport.ScharfetterGummelGraded

ChargeTransport.SchottkyContact

ChargeTransport.StandardFuncSet

ChargeTransport.Stationary

ChargeTransport.System

ChargeTransport.System

ChargeTransport.Transient

ChargeTransport.Traps

Methods

ChargeTransport.Blakemore

ChargeTransport.Boltzmann

ChargeTransport.FermiDiracMinusOne

ChargeTransport.FermiDiracOneHalfBednarczyk

ChargeTransport.FermiDiracOneHalfTeSCA

ChargeTransport.RHSContinuityEquations!

ChargeTransport.RHSPoisson!

ChargeTransport.bflux!

ChargeTransport.breaction!

ChargeTransport.breaction!

ChargeTransport.breaction!

ChargeTransport.bstorage!

ChargeTransport.charge_density

ChargeTransport.charge_density

ChargeTransport.compute_densities!

ChargeTransport.compute_densities!

ChargeTransport.compute_energies!

ChargeTransport.degenerateLimit

ChargeTransport.electroNeutralSolution!

ChargeTransport.electrochemicalReaction

ChargeTransport.enable_ionic_carriers

ChargeTransport.enable_traps!

ChargeTransport.etaFunction

ChargeTransport.etaFunction

ChargeTransport.etaFunction

ChargeTransport.etaFunction

ChargeTransport.flux!

ChargeTransport.get_BEE!

ChargeTransport.get_DOS!

ChargeTransport.get_current_val

ChargeTransport.get_current_val

ChargeTransport.plotDiffusionEnhancements

ChargeTransport.plotDistributions

ChargeTransport.plot_IV

ChargeTransport.plot_densities

ChargeTransport.plot_doping

ChargeTransport.plot_doping

ChargeTransport.plot_electroNeutralSolutionBoltzmann

ChargeTransport.plot_energies

ChargeTransport.plot_energies

ChargeTransport.plot_solution

ChargeTransport.printJacobi

ChargeTransport.reaction!

ChargeTransport.reaction!

ChargeTransport.reaction!

ChargeTransport.set_bulk_recombination

ChargeTransport.set_plotting_labels

ChargeTransport.storage!

ChargeTransport.storage!

ChargeTransport.trap_density!

Description of Constant, Types, Constructors and Methods

Constants, Types and Constructors

ChargeTransport.BoundaryModelType — Type

Possible types of boundary models.

ChargeTransport.CalculationType — Type

Possible types for calculation type.

ChargeTransport.FluxApproximationType — Type

Possible types of flux discretization schemes.

ChargeTransport.GenerationModelType — Type

Possible types for generation model.

ChargeTransport.InterfaceModelType — Type

Possible Types of interface model (interior boundary conditions).

ChargeTransport.ModelType — Type

Possible types which indicate, if we consider stationary or transient problem.

ChargeTransport.OuterBoundaryModelType — Type

Possible types of outer boundary model.

ChargeTransport.QType — Type

Type of charge carriers and the electric potential.

ChargeTransport.SRHWithoutTrapsType — Type

Possible type for SRH recombination without traps.

ChargeTransport.StandardFuncSet — Type

Type of statistics functions.

ChargeTransport.BulkRecombination — Type

mutable struct BulkRecombination

A struct holding all necessary information for building bulk recombination. With help of this constructor we can read out the indices the user chooses for electron and hole quasi Fermi potentials.

iphin::Int64

Index for FVM construction of electron quasi Fermi potential.

iphip::Int64

Index for FVM construction of hole quasi Fermi potential.

bulk_recomb_Auger::Bool

Boolean for present Auger recombination in bulk.

bulk_recomb_radiative::Bool

Boolean for present radiative recombination in bulk.

bulk_recomb_SRH::Union{Type{SRHOff}, Type{ChargeTransport.SRHStationary}, Type{SRHTrapsTransient}}

DataType for present SRH recombination in bulk. This needs to be a Type due to cases with or without mobile traps.

ChargeTransport.Data — Type

mutable struct Data{TFuncs<:Function}

A struct holding all data information including model and numerics information, but also all physical parameters for a drift-diffusion simulation of a semiconductor device.

F::Vector{TFuncs} where TFuncs<:Function

An array with the corresponding distribution function $\mathcal{F}_\alpha$ for all carriers $\alpha$.

boundaryType::Vector{Union{Type{InterfaceModelDiscontqF}, Type{InterfaceModelIonCharge}, Type{InterfaceModelNone}, Type{InterfaceModelSurfaceReco}, Type{InterfaceModelSurfaceRecoAndTangentialFlux}, Type{InterfaceModelTangentialFlux}, Type{OhmicContact}, Type{SchottkyContact}}}

An array of DataTypes with the type of boundary model for each boundary (interior and exterior).

bulkRecombination::BulkRecombination

A struct containing information concerning the bulk recombination model.

enableIonicCarriers::IonicChargeCarriers

A struct which contains information on the regions, where ionic charge carriers (ions and/or ion vacancies) are present.

enableTraps::ChargeTransport.Traps

A struct which contains information on present SRH traps.

innerInterfaceModel::Union{Type{InterfaceModelDiscontqF}, Type{InterfaceModelIonCharge}, Type{InterfaceModelNone}, Type{InterfaceModelSurfaceReco}, Type{InterfaceModelSurfaceRecoAndTangentialFlux}, Type{InterfaceModelTangentialFlux}}

DataType which stores information about which inner interface model is chosen by user. This quantity cannot be seen by the user and is needed for the core of the package.

fluxApproximation::Union{Type{DiffusionEnhanced}, Type{ExcessChemicalPotential}, Type{ExcessChemicalPotentialGraded}, Type{GeneralizedSG}, Type{ScharfetterGummel}, Type{ScharfetterGummelGraded}}

A DataType for the flux discretization method.

calculationType::Union{Type{InEquilibrium}, Type{OutOfEquilibrium}}

A DataType for equilibrium or out of equilibrium calculations.

modelType::Union{Type{Stationary}, Type{Transient}}

A DataType for transient or stationary calculations.

generationModel::Union{Type{GenerationBeerLambert}, Type{GenerationNone}, Type{GenerationUniform}}

A DataType for for generation model.

λ1::Float64

An embedding parameter used to solve the nonlinear Poisson problem, where for λ1 = 0 the right hand-side is set to zero whereas for for λ1 = 1 we have a full space charge density.

λ2::Float64

An embedding parameter for the generation rate.

λ3::Float64

An embedding parameter for an electrochemical reaction.

tempBEE1::Vector{Float64}

Within this template informations concerning the band-edge energy of each carrier is stored locally which saves allocations. We have two of such templates due to the two point flux approximation schemes.

tempBEE2::Vector{Float64}

See the description of tempBEE1.

tempDOS1::Vector{Float64}

Within this template informations concerning the effective DOS of each carrier is stored locally which saves allocations. We have two of such templates due to the two point flux approximation schemes.

tempDOS2::Vector{Float64}

See the desciption of tempDOS2.

isContinuous::Vector{Bool}

An array containing information on whether charge carriers are continuous or discontinuous. This is needed for building the AbstractQuantities which handle the indices of charge carriers on different regions.

chargeCarrierList::Vector{Union{Int64, VoronoiFVM.ContinuousQuantity{Int64}, VoronoiFVM.DiscontinuousQuantity{Int64}}}

This list stores all charge carriers. Based on the user choice we have with this new type the opportunity to simulate discontinuous unknowns.

index_psi::Union{Int64, VoronoiFVM.ContinuousQuantity{Int64}, VoronoiFVM.DiscontinuousQuantity{Int64}}

This variable stores the index of the electric potential. Based on the user choice we have with this new type the opportunity to simulate discontinuous unknowns.

params::Params

A struct holding all region dependent parameter information. For more information see struct Params.

paramsnodal::ParamsNodal

A struct holding all space dependent parameter information. For more information see struct ParamsNodal.

ChargeTransport.Data — Method

Data(grid, numberOfCarriers; statfunctions) -> Data{StandardFuncSet}

Simplified constructor for Data which only takes the grid and the numberOfCarriers as argument. Here, all necessary information including the physical parameters, but also some numerical information are located.

ChargeTransport.DiffusionEnhanced — Type

abstract type DiffusionEnhanced

Abstract type for diffusion enhanced flux discretization, check M. Bessemoulin-Chatard, “A finite volume scheme for convection–diffusion equations with nonlinear diffusion derived from the Scharfetter–Gummel scheme”, Numerische Mathematik, vol. 121, pp. 637–670, 2012.

ChargeTransport.ExcessChemicalPotential — Type

abstract type ExcessChemicalPotential

Abstract type for excess chemical potential flux discretization, check Z. Yu, and R. Dutton, “SEDAN III – A one-dimensional device simulator”, http://www-tcad.stanford.edu/tcad/programs/sedan3.html, 1988.

ChargeTransport.ExcessChemicalPotentialGraded — Type

abstract type ExcessChemicalPotentialGraded

Abstract type for excess chemical potential flux discretization for graded effective density of states and/or graded band-edge energies. This means, use this flux when at least one of these parameters is assumed to be space-dependent.

ChargeTransport.GeneralizedSG — Type

abstract type GeneralizedSG

Abstract type for generalized Scharfetter-Gummel flux discretization. This flux approximation results in an implicit equation which needs to be solved and is exact for all Blakemore type statistics functions with abritary γ, check T. Koprucki and K. Gärtner. “Discretization scheme for drift-diffusion equations with strong diffusion enhancement”. In: 12th International Conference on Numerical Simulation of Optoelectronic Devices (NUSOD). 2012, pp. 103–104.

ChargeTransport.GenerationBeerLambert — Type

abstract type GenerationBeerLambert

Abstract type for Beer-Lambert generation. Note that this type is implemented, but not well tested yet.

ChargeTransport.GenerationNone — Type

abstract type GenerationNone

Abstract type for no generation model.

ChargeTransport.GenerationUniform — Type

abstract type GenerationUniform

Abstract type for uniform generation.

ChargeTransport.InEquilibrium — Type

abstract type InEquilibrium

Abstract type for equilibrium calculations.

ChargeTransport.InterfaceModelNone — Type

abstract type InterfaceModelNone

Abstract type for no interface model.

ChargeTransport.InterfaceModelSurfaceReco — Type

abstract type InterfaceModelSurfaceReco

Abstract type for surface recombination mechanisms.

ChargeTransport.IonicChargeCarriers — Type

mutable struct IonicChargeCarriers

A struct holding all information necessary on the ionic charge carriers. With help of this constructor we can read out the indices the user chooses for ionic charge carrier quasi Fermi potentials and the respective regions in which they are defined. Note that it is possible to use ions as well as ion vacancies.

ionic_carriers::Vector{Int64}

Array with the indices of ionic charge carriers.

regions::Vector{Int64}

Corresponding regions where ionic charge carriers are assumed to be present.

ChargeTransport.OhmicContact — Type

Abstract type for ohmic contacts as outer boundary model.

ChargeTransport.OutOfEquilibrium — Type

abstract type OutOfEquilibrium

Abstract type for out of equilibrium calculations.

ChargeTransport.Params — Type

mutable struct Params

A struct holding the physical region dependent parameters for a drift-diffusion simulation of a semiconductor device.

numberOfNodes::Int64

Number of nodes used for the disretization of the domain $\mathbf{\Omega}$.

numberOfRegions::Int64

Number of subregions $\mathbf{\Omega}_k$ within the domain $\mathbf{\Omega}$.

numberOfBoundaryRegions::Int64

Number of boundary regions $(\partial \mathbf{\Omega})_k$ such that $\partial \mathbf{\Omega} = \cup_k (\partial \mathbf{\Omega})_k$. Note that here are interior and outerior boundaries calculated.

numberOfCarriers::Int64

Number of moving charge carriers.

temperature::Float64

A given constant temperature.

UT::Float64

The thermal voltage, which reads $U_T = k_B T / q$.

γ::Float64

The parameter of the Blakemore statistics (needed for the generalizedSG flux).

r0::Float64

Prefactor of electro-chemical reaction of internal boundary conditions.

prefactor_SRH::Float64

Prefactor for stationary SRH recombination.

SchottkyBarrier::Vector{Float64}

An array for the given Schottky barriers at present Schotkky contacts.

chargeNumbers::Vector{Float64}

An array with the corresponding charge numbers $z_\alpha$ for all carriers $\alpha$.

bBandEdgeEnergy::Matrix{Float64}

An array with the corresponding boundary band-edge energy values $E_\alpha$ in each region for each carrier $\alpha$.

bDensityOfStates::Matrix{Float64}

An array with the corresponding boundary effective density of states values $N_\alpha$ for each carrier $\alpha$.

bMobility::Matrix{Float64}

A 2D array with the corresponding boundary mobility values $\mu_\alpha$ in each boundary region for each carrier $\alpha$.

bDoping::Matrix{Float64}

A 2D array with the corresponding boundary doping values for each carrier $\alpha$.

bVelocity::Matrix{Float64}

A 2D array with the corresponding boundary velocity values for each carrier $\alpha$, when assuming Schottky contacts.

recombinationSRHvelocity::Matrix{Float64}

A 2D array with the corresponding recombination surface boundary velocity values for electrons and holes.

bRecombinationSRHTrapDensity::Matrix{Float64}

A 2D array with the corresponding recombination surface boundary density values for electrons and holes.

doping::Matrix{Float64}

A 2D array with the corresponding doping values for each carrier $\alpha$ on each region.

densityOfStates::Matrix{Float64}

A 2D array with the corresponding effective density of states values $N_\alpha$ for each carrier $\alpha$ on each region.

bandEdgeEnergy::Matrix{Float64}

A 2D array with the corresponding band-edge energy values $E_\alpha$ for each carrier $\alpha$ on each region.

mobility::Matrix{Float64}

A 2D array with the corresponding mobility values $\mu_\alpha$ for each carrier $\alpha$ on each region.

recombinationSRHLifetime::Matrix{Float64}

A 2D array with the corresponding SRH lifetimes $\tau_n, \tau_p$ for electrons and holes.

recombinationSRHTrapDensity::Matrix{Float64}

A 2D array with the corresponding time-independent SRH trap densities $n_{\tau}, p_{\tau}$ for electrons and holes.

recombinationAuger::Matrix{Float64}

A 2D array with the corresponding Auger coefficients for electrons and holes.

dielectricConstant::Vector{Float64}

A region dependent dielectric constant.

generationIncidentPhotonFlux::Vector{Float64}

A region dependent array for the prefactor in the generation process which is the incident photon flux.

generationUniform::Vector{Float64}

A region dependent array for an uniform generation rate.

generationAbsorption::Vector{Float64}

A region dependent array for the absorption coefficient in the generation process.

recombinationRadiative::Vector{Float64}

A region dependent array for the radiative recombination rate.

ChargeTransport.Params — Method

Params(grid, numberOfCarriers) -> Params

Simplified constructor for Params which only takes the grid and the numberOfCarriers as argument.

ChargeTransport.ParamsNodal — Type

mutable struct ParamsNodal

A struct holding the physical nodal, i.e. space-dependent parameters for a drift-diffusion simulation of a semiconductor device.

dielectricConstant::Vector{Float64}

A node dependent dielectric constant.

doping::Vector{Float64}

A 1D array with the corresponding doping values on each node.

mobility::Matrix{Float64}

A 2D array with the corresponding mobility values $\mu_\alpha$ for each carrier $\alpha$ on each node.

densityOfStates::Matrix{Float64}

A 2D array with the corresponding effective density of states values $N_\alpha$ for each carrier $\alpha$ on each node.

bandEdgeEnergy::Matrix{Float64}

A 2D array with the corresponding band-edge energy values $E_\alpha$ for each carrier $\alpha$ on each node.

ChargeTransport.ParamsNodal — Method

ParamsNodal(grid, numberOfCarriers) -> ParamsNodal

Simplified constructor for ParamsNodal which only takes the grid and the numberOfCarriers as argument.

ChargeTransport.SRHWithTrapsType — Type

Possible types for SRH recombination without traps.

ChargeTransport.ScharfetterGummel — Type

abstract type ScharfetterGummel

Abstract type for Scharfetter-Gummel flux discretization. Choose this one, when the Boltzmann statistics function is chosen as statistics, check D. Scharfetter and H. Gummel, “Large-signal analysis of a silicon Read diode oscillator”, IEEE Trans. Electr. Dev., vol. 16, pp. 64–77, 1969.

ChargeTransport.ScharfetterGummelGraded — Type

abstract type ScharfetterGummelGraded

Abstract type for Scharfetter-Gummel flux discretization for graded effective density of states and/or graded band-edge energies. This means, use this flux when at least one of these parameters is assumed to be space-dependent.

ChargeTransport.SchottkyContact — Type

Abstract type for schottky contacts as boundary model.

ChargeTransport.Stationary — Type

abstract type Stationary

Abstract type for stationary simulations.

ChargeTransport.System — Type

mutable struct System

A struct holding all information necessary for a drift-diffusion type system.

data::Data

A struct holding all data information, see Data

fvmsys::VoronoiFVM.AbstractSystem

A struct holding system information for the finite volume system.

ChargeTransport.System — Method

System(grid, data; unknown_storage)

System constructor which builds all necessary information needed based on the input parameters with special regard to interface models. This is the main struct in which all information on the input data, but also on the solving system, are stored.

ChargeTransport.Transient — Type

abstract type Transient

Abstract type for transient simulations.

ChargeTransport.Traps — Type

mutable struct Traps

A struct holding all information necessary for enabling traps in the SRH recombination. With help of this constructor we can read out the index the user chooses for trap quasi Fermi potentials and the respective regions in which they are defined.

traps::Int64

Array with the index of traps.

regions::Vector{Int64}

Corresponding regions where traps are assumed to be present.

Methods

ChargeTransport.Blakemore — Method

Blakemore(x::Real, γ::Real) -> Any

The Blakemore approximation $1/(\exp(-x) + γ)$ with variable real scalar $γ$, see J. S. Blakemore. “The Parameters of Partially Degenerate Semiconductors”. In: Proceedings of the Physical Society. Section A 65 (1952), pp. 460–461.

ChargeTransport.Boltzmann — Method

Boltzmann(x::Real) -> Any

The Boltzmann statistics function $\exp(x)$.

ChargeTransport.FermiDiracMinusOne — Method

FermiDiracMinusOne(x::Real) -> Any

The Fermi-Dirac integral of order $-1$ which reads $1/(\exp(-x) + 1)$, see J.S. Blakemore, Approximations for Fermi-Dirac integrals, especially the function $F_{1/2} (\eta)$ used to describe electron density in a semiconductor, Solid-State Electronics 25 (11) (1982) 1067 – 1076.

ChargeTransport.FermiDiracOneHalfBednarczyk — Method

FermiDiracOneHalfBednarczyk(x::Real) -> Any

The incomplete Fermi-Dirac integral of order 1/2, implemented according to [Bednarczyk1978, "The Approximation of the Fermi-Dirac integral $F_{1/2}(\eta)$"].

ChargeTransport.FermiDiracOneHalfTeSCA — Method

FermiDiracOneHalfTeSCA(x::Real) -> Any

The incomplete Fermi-Dirac integral of order 1/2, implemented according to the software package TeSCA, see https://wias-berlin.de/software/index.jsp?lang=1&id=TeSCA.

ChargeTransport.RHSContinuityEquations! — Method

RHSContinuityEquations!(f, u, node, data)

Function which builds right-hand side of electric charge carriers.

ChargeTransport.RHSPoisson! — Method

RHSPoisson!(f, u, node, data) -> Any

Function which builds right-hand side of Poisson equation, i.e. which builds the space charge density.

ChargeTransport.bflux! — Method

bflux!(f, u, bedge, data) -> Any

Master bflux! function. This is the function which enters VoronoiFVM and hands over for each boundary the flux within the boundary.

ChargeTransport.breaction! — Method

breaction!(f, u, bnode, data) -> Any

Master breaction! function. This is the function which enters VoronoiFVM and hands over for each boundary the chosen boundary model.

ChargeTransport.breaction! — Method

breaction!(f, u, bnode, data, _::Type{OhmicContact}) -> Any

Creates ohmic boundary conditions via a penalty approach with penalty parameter $\delta$. For example, the right-hand side for the electrostatic potential $\psi$ is implemented as

$f[\psi] = -q/\delta ( (p - N_a) - (n - N_d) )$,

assuming a bipolar semiconductor. In general, we have for some given charge number $z_\alpha$

$f[\psi] = -q/\delta \sum_\alpha{ z_\alpha (n_\alpha - C_\alpha) },$

where $C_\alpha$ corresponds to some doping w.r.t. the species $\alpha$.

The boundary conditions for the charge carriers are set in the main file. Hence,

$f[n_\alpha] = 0$`

for all charge carriers $\alpha$.

ChargeTransport.breaction! — Method

breaction!(f, u, bnode, data, _::Type{SchottkyContact})

Creates Schottky boundary conditions in a first attempt. For the electrostatic potential we assume

$\psi = \phi_S/q + U,$

where $\phi_S$ corresponds to a given value (Schottky barrier) and $U$ to the applied voltage. For now, the quantitity $\phi_S$ needs to be specified in the main file. For the charge carriers we assume the following

$f[n_\alpha] = z_\alpha q v_\alpha (n_\alpha - n_{\alpha, 0})$,

where $v_{\alpha}$ can be treated as a surface recombination mechanism and is given. The parameter $n_{\alpha, 0}$ is a given value, calculated by the statistical relation, when assuming no electrical field and a quasi Fermi level equal to the Schottky barrier $\phi_S$, i.e.

$n_{\alpha, 0}= z_\alpha/ U_T (E_\alpha - \phi_S) / q.$

ChargeTransport.bstorage! — Method

bstorage!(f, u, bnode, data) -> Any

Master bstorage! function. This is the function which enters VoronoiFVM and hands over for each boundary the time-dependent part of the chosen boundary model.

ChargeTransport.charge_density — Method

charge_density(psi0, phi, UT, EVector, chargeNumbers, dopingVector, dosVector, FVector) -> Any

Compute the charge density, i.e. the right-hand side of Poisson's equation.

ChargeTransport.charge_density — Method

charge_density(ctsys, sol) -> Any

Compute the charge density for each region separately.

ChargeTransport.compute_densities! — Method

compute_densities!(u, data, inode, region, icc, in_region::Bool) -> Any

For given potentials, compute corresponding densities. This function is needed for the method, plotting the densities.

ChargeTransport.compute_densities! — Method

compute_densities!(grid, data, sol) -> Matrix{Real}

For given potentials in vector form, compute corresponding vectorized densities. [Caution: this was not tested for multidimensions.]

ChargeTransport.compute_energies! — Method

compute_energies!(grid, data, sol)

For given solution in vector form, compute corresponding vectorized band-edge energies and Fermi level. [Caution: this was not tested for multidimensions.]

ChargeTransport.degenerateLimit — Method

degenerateLimit(x) -> Any

Degenerate limit of incomplete Fermi-Dirac integral of order 1/2.

ChargeTransport.electroNeutralSolution! — Method

electroNeutralSolution!(grid, data; Newton) -> Any

Compute the electro-neutral solution for the Boltzmann approximation. It is obtained by setting the left-hand side in the Poisson equation equal to zero and solving for $\psi$. The charge carriers may obey different statitics functions. Currently, this one is not well tested for the case of charge carriers beyond electrons and holes.

ChargeTransport.electrochemicalReaction — Method

electrochemicalReaction(data, u, iphia, ipsi, iphiaJunction, ipsiJunction, β, κ, DOS, E) -> Any

Electrochemical reaction between interface and bulk ionic species. This function enters in the internal boundary reaction in case of an ion charge interface model.

ChargeTransport.enable_ionic_carriers — Method

Corresponding constructor for the present ionic charge carriers and the respective regions.

ChargeTransport.enable_traps! — Method

Corresponding constructor for the present traps and the respective regions.

ChargeTransport.etaFunction — Method

etaFunction(psi, phi, UT, E, z) -> Any

The argument of the distribution function for given $\varphi_\alpha$ and $\psi$

$z_\alpha / U_T ( (\varphi_\alpha - \psi) + E_\alpha / q ).$

The parameters $E_\alpha$ and $z_\alpha$ are given as vectors. This function may be used to compute the charge density, i.e. the right-hand side of the Poisson equation.

ChargeTransport.etaFunction — Method

etaFunction(u, data, node, region, icc, in_region::Bool) -> Any

The argument of the distribution function for floats.

ChargeTransport.etaFunction — Method

etaFunction(u, bnode::VoronoiFVM.BNode, data, icc) -> Any

The argument of the distribution function for boundary nodes.

ChargeTransport.etaFunction — Method

etaFunction(u, node::VoronoiFVM.Node, data, icc) -> Any

The argument of the distribution function for interior nodes.

ChargeTransport.flux! — Method

flux!(f, u, edge, data) -> Any

Master flux functions which enters VoronoiFVM. Flux discretization scheme is chosen in two steps. First, we need to see, if we are in or out of equilibrium. If, InEquilibrium, then no flux is passed. If outOfEquilibrium, we choose the flux approximation which the user chose.

ChargeTransport.get_BEE! — Method

get_BEE!(icc::Union{Int64, VoronoiFVM.ContinuousQuantity{Int64}, VoronoiFVM.DiscontinuousQuantity{Int64}}, node::VoronoiFVM.Node, data) -> Any

Defining locally the band-edge energy for interior nodes (analougesly for boundary nodes and edges).

ChargeTransport.get_DOS! — Method

get_DOS!(icc::Union{Int64, VoronoiFVM.ContinuousQuantity{Int64}, VoronoiFVM.DiscontinuousQuantity{Int64}}, node::VoronoiFVM.Node, data) -> Any

Defining locally the effective DOS for interior nodes (analogously for boundary nodes and edges).

ChargeTransport.get_current_val — Method

Calculates current for time dependent problem.

ChargeTransport.get_current_val — Method

Calculates current for stationary problem.

ChargeTransport.plotDiffusionEnhancements — Method

plotDiffusionEnhancements()

Plot diffusion enhancements.

ChargeTransport.plotDistributions — Method

plotDistributions(; Plotter)

Plot different distribution integrals.

ChargeTransport.plot_IV — Method

plot_IV(Plotter, biasValues, IV, title; plotGridpoints) -> Any

Method for showing the total current. One input parameter is the boolean plotGridpoints which makes it possible to plot markers, which indicate where the nodes are located.

ChargeTransport.plot_densities — Method

plot_densities(Plotter, grid, data::Data, sol, title, label_density; plotGridpoints) -> Any

Plotting routine, where the charge carrier densities are depicted in dependence of space. The case of heterojunctions is tested, but yet multidimensional plottings are not included. One input parameter is the boolean plotGridpoints which makes it possible to plot markers, which indicate where the nodes are located.

ChargeTransport.plot_doping — Method

plot_doping(Plotter, g::ExtendableGrids.ExtendableGrid, data::Data, label_density) -> Any

Possibility to plot the considered doping. This is especially useful for making sure that the interior and the boundary doping agree.

ChargeTransport.plot_doping — Method

Plot doping for nodal dependent doping.

ChargeTransport.plot_electroNeutralSolutionBoltzmann — Method

plot_electroNeutralSolutionBoltzmann(Plotter, grid, psi0; plotGridpoints) -> Any

Plotting routine for depicting the electroneutral potential. One input parameter is the boolean plotGridpoints which makes it possible to plot markers, which indicate where the nodes are located.

ChargeTransport.plot_energies — Method

plot_energies(Plotter, grid, data::Data, sol, title, label_energy; plotGridpoints) -> Any

With this method it is possible to plot the energies

$E_\alpha - q \psi \quad \text{w.r.t. space.}$

The case of heterojunctions is tested, but yet multidimensional plottings are not included.

One input parameter is the boolean plotGridpoints which makes it possible to plot markers, which indicate where the nodes are located.

ChargeTransport.plot_energies — Method

plot_energies(Plotter, grid, data, label_BEE)

With this method it is possible to depict the band-edge energies $E_\alpha$. This can be useful for debugging when dealing with heterojunctions.

ChargeTransport.plot_solution — Method

plot_solution(Plotter, grid, data::Data, solution, title, label_solution; plotGridpoints) -> Any

Method for plotting the solution vectors: the electrostatic potential $\psi$ as well as the charge carriers. The case of heterojunctions is tested, but yet multidimensional plottings are not included. One input parameter is the boolean plotGridpoints which makes it possible to plot markers, which indicate where the nodes are located.

ChargeTransport.printJacobi — Method

printJacobi(node, sys)

First try of debugger. Print the Jacobi matrix for a given node, i.e. the number of node in the grid and not the excact coordinate. This is only done for the one dimensional case so far.

ChargeTransport.reaction! — Method

reaction!(f, u, node, data)

Master reaction! function. This is the function which enters VoronoiFVM and hands over reaction terms for concrete calculation type and bulk recombination model.

ChargeTransport.reaction! — Method

reaction!(f, u, node, data, _::Type{InEquilibrium})

Reaction in case of equilibrium, i.e. no generation and recombination is considered.

ChargeTransport.reaction! — Method

reaction!(f, u, node, data, _::Type{OutOfEquilibrium})

Sets up the right-hand sides. Assuming a bipolar semiconductor the right-hand side for the electrostatic potential becomes

$f[ψ] = - q ((p - N_a) - (n - N_d) ) = - q \sum n_\alpha (n_\alpha - C_\alpha)$

for some doping $C_\alpha$ w.r.t. to the species $\alpha$. The right-hand sides for the charge carriers read as

$f[n_\alpha] = - z_\alpha q (G - R)$

for all charge carriers $n_\alpha$. The recombination includes radiative, Auger and Shockley-Read-Hall recombination. For latter recombination process the stationary simplification is implemented.

The recombination is only implemented for electron and holes and assumes that the electron index is 1 and the hole index is 2.

ChargeTransport.set_bulk_recombination — Method

set_bulk_recombination(; iphin, iphip, bulk_recomb_Auger, bulk_recomb_radiative, bulk_recomb_SRH)

Corresponding constructor for the bulk recombination model.

ChargeTransport.set_plotting_labels — Method

set_plotting_labels(data) -> Tuple{Any, Any, Matrix{String}, Any}

Method which can be used to construct the arrays parsed to the plotting routines for labeling. The description for electrons and holes are predefined. If one wishes to extend by labels for, e.g. mobile ionic carriers or traps, this can be done within the main file.

ChargeTransport.storage! — Method

storage!(f, u, node, data) -> Union{Nothing, Float64}

Master storage! function. This is the function which enters VoronoiFVM and hands over a storage term, if we consider transient problem.

ChargeTransport.storage! — Method

storage!(f, u, node, data, _::Type{Transient}) -> Float64

The storage term for time-dependent problems. Currently, for the time-dependent current densities the implicit Euler scheme is used. Hence, we have

$f[n_\alpha] = z_\alpha q ∂_t n_\alpha$

and for the electrostatic potential $f[ψ] = 0$.

ChargeTransport.trap_density! — Method

trap_density!(icc, ireg, data, Et)

Compute trap densities for a given trap energy. [Currently, only done for the Boltzmann statistics and for region dependent parameters.]