Overview – Types, Constructors and Methods

Types and Constructors

ChargeTransportInSolids.ChargeTransportBulkRecombination

ChargeTransportInSolids.ChargeTransportData

ChargeTransportInSolids.ChargeTransportData

ChargeTransportInSolids.ChargeTransportIonicChargeCarriers

ChargeTransportInSolids.ChargeTransportParams

ChargeTransportInSolids.ChargeTransportParams

ChargeTransportInSolids.ChargeTransportParamsNodal

ChargeTransportInSolids.ChargeTransportParamsNodal

ChargeTransportInSolids.ChargeTransportSystem

ChargeTransportInSolids.ChargeTransportSystem

ChargeTransportInSolids.ChargeTransportTraps

ChargeTransportInSolids.abstract_model_SRH

ChargeTransportInSolids.boundary_model

ChargeTransportInSolids.calculation_type

ChargeTransportInSolids.diffusion_enhanced

ChargeTransportInSolids.excess_chemical_potential

ChargeTransportInSolids.excess_chemical_potential_graded

ChargeTransportInSolids.flux_approximation

ChargeTransportInSolids.generalized_sg

ChargeTransportInSolids.generation_beer_lambert

ChargeTransportInSolids.generation_model

ChargeTransportInSolids.generation_none

ChargeTransportInSolids.generation_uniform

ChargeTransportInSolids.inEquilibrium

ChargeTransportInSolids.interface_model

ChargeTransportInSolids.interface_model_discont_qF

ChargeTransportInSolids.interface_model_ion_charge

ChargeTransportInSolids.interface_model_none

ChargeTransportInSolids.interface_model_surface_recombination

ChargeTransportInSolids.linearScanProtocol

ChargeTransportInSolids.model_SRH

ChargeTransportInSolids.model_SRH_2species_present_trap_dens

ChargeTransportInSolids.model_SRH_with_traps

ChargeTransportInSolids.model_SRH_without_traps

ChargeTransportInSolids.model_stationary

ChargeTransportInSolids.model_transient

ChargeTransportInSolids.model_type

ChargeTransportInSolids.ohmic_contact

ChargeTransportInSolids.outOfEquilibrium

ChargeTransportInSolids.scan_protocol_type

ChargeTransportInSolids.scharfetter_gummel

ChargeTransportInSolids.scharfetter_gummel_graded

ChargeTransportInSolids.schottky_contact

Methods

ChargeTransportInSolids.Blakemore

ChargeTransportInSolids.Blakemore

ChargeTransportInSolids.Boltzmann

ChargeTransportInSolids.FermiDiracMinusOne

ChargeTransportInSolids.FermiDiracOneHalfBednarczyk

ChargeTransportInSolids.FermiDiracOneHalfTeSCA

ChargeTransportInSolids.addChargeCarriers!

ChargeTransportInSolids.addElectricPotential!

ChargeTransportInSolids.addTrapDensity!

ChargeTransportInSolids.bflux!

ChargeTransportInSolids.bflux!

ChargeTransportInSolids.breaction!

ChargeTransportInSolids.breaction!

ChargeTransportInSolids.breaction!

ChargeTransportInSolids.breaction!

ChargeTransportInSolids.bstorage!

ChargeTransportInSolids.bstorage!

ChargeTransportInSolids.bstorage!

ChargeTransportInSolids.build_system

ChargeTransportInSolids.build_system

ChargeTransportInSolids.chargeDensity

ChargeTransportInSolids.chargeDensity

ChargeTransportInSolids.compute_densities!

ChargeTransportInSolids.compute_densities!

ChargeTransportInSolids.compute_energies!

ChargeTransportInSolids.degenerateLimit

ChargeTransportInSolids.electroNeutralSolution!

ChargeTransportInSolids.electrochemicalReaction

ChargeTransportInSolids.enable_ionic_carriers

ChargeTransportInSolids.enable_traps!

ChargeTransportInSolids.enable_traps!

ChargeTransportInSolids.etaFunction

ChargeTransportInSolids.etaFunction

ChargeTransportInSolids.etaFunction

ChargeTransportInSolids.etaFunction

ChargeTransportInSolids.flux!

ChargeTransportInSolids.flux!

ChargeTransportInSolids.flux!

ChargeTransportInSolids.flux!

ChargeTransportInSolids.flux!

ChargeTransportInSolids.flux!

ChargeTransportInSolids.generation

ChargeTransportInSolids.generation

ChargeTransportInSolids.get_BEE!

ChargeTransportInSolids.get_BEE!

ChargeTransportInSolids.get_BEE!

ChargeTransportInSolids.get_DOS!

ChargeTransportInSolids.get_DOS!

ChargeTransportInSolids.get_DOS!

ChargeTransportInSolids.get_current_val

ChargeTransportInSolids.get_current_val

ChargeTransportInSolids.inner_interface_model

ChargeTransportInSolids.inner_interface_model

ChargeTransportInSolids.kernelSRH

ChargeTransportInSolids.kernelSRH

ChargeTransportInSolids.plotDiffusionEnhancements

ChargeTransportInSolids.plotDistributions

ChargeTransportInSolids.plot_IV

ChargeTransportInSolids.plot_densities

ChargeTransportInSolids.plot_doping

ChargeTransportInSolids.plot_doping

ChargeTransportInSolids.plot_electroNeutralSolutionBoltzmann

ChargeTransportInSolids.plot_energies

ChargeTransportInSolids.plot_energies

ChargeTransportInSolids.plot_solution

ChargeTransportInSolids.printJacobi

ChargeTransportInSolids.reaction!

ChargeTransportInSolids.reaction!

ChargeTransportInSolids.reaction!

ChargeTransportInSolids.set_bulk_recombination

ChargeTransportInSolids.set_time_mesh

ChargeTransportInSolids.storage!

ChargeTransportInSolids.storage!

ChargeTransportInSolids.trap_density!

Description of Types, Constructors and Methods

Types and Constructors

ChargeTransportInSolids.ChargeTransportBulkRecombination — Type

mutable struct ChargeTransportBulkRecombination

A struct holding all information necessary for building bulk recombination. With help of this constructor we can read out the indices the user chooses for electron and hole quasi Fermi potentials.

iphin::Int64

index for data construction of quasi Fermi potential of electrons

iphip::Int64

index for data construction of quasi Fermi potential of holes

bulk_recomb_Auger::Bool

boolean for present Auger recombination in bulk

bulk_recomb_radiative::Bool

boolean for present radiative recombination in bulk

bulk_recomb_SRH::DataType

DataType for present SRH recombination in bulk

model_SRH_2species_trap::DataType

Auxiliary quantitiy for simulations with present trap density in Poisson, but without traps as own unknowns. Note that this one may be deleted in future versions.

ChargeTransportInSolids.ChargeTransportData — Type

mutable struct ChargeTransportData

A struct holding all data information including model and numerics information, but also all physical parameters for a drift-diffusion simulation of a semiconductor device.

F::Vector{Function}

An array with the corresponding distribution function $\mathcal{F}_\alpha$ for all carriers $\alpha$.

boundary_type::Vector{DataType}

An array of DataTypes with the type of boundary model for each boundary (interior and exterior).

bulk_recombination::ChargeTransportBulkRecombination

A struct containing information concerning the bulk recombination model.

enable_ionic_carriers::ChargeTransportIonicChargeCarriers

A struct which contains information on the regions, where ionic charge carriers (ions and/or ion vacancies) are present.

enable_traps::ChargeTransportInSolids.ChargeTransportTraps

An AbstractVector which contains information on present SRH traps.

inner_interface_model::DataType

DataType which stores information about which inner interface model is chosen by user. This quantity cannot be seen by the user and is needed for the core of package.

flux_approximation::DataType

A DataType for the flux discretization method.

calculation_type::DataType

A DataType for equilibrium or out of equilibrium calculations.

model_type::DataType

A DataType for transient or stationary calculations.

generation_model::DataType

A DataType for for generation model.

λ1::Float64

An embedding parameter used to solve the nonlinear Poisson problem, which corresponds for λ1 = 0 to the case of thermodynamic equilibrium and electrocharge neutrality and for λ1 = 1 to the case of thermodynamic equilibrium.

λ2::Float64

An embedding parameter for turning the generation rate $G$ on.

λ3::Float64

An embedding parameter for electrochemical reaction.

tempBEE1::Vector{Float64}

Within this template informations concerning the band-edge energy of each carrier is stored locally which saves allocations. We have two of such templates due to the two point flux approximation schemes.

tempBEE2::Vector{Float64}

see the description of tempBEE1

tempDOS1::Vector{Float64}

Within this template informations concerning the effective DOS of each carrier is stored locally which saves allocations. We have two of such templates due to the two point flux approximation schemes.

tempDOS2::Vector{Float64}

see the desciption of tempDOS2

isContinuous::Vector{Bool}

An array which contains information on whether charge carriers are continuous or discontinuous. This is needed for building the AbstractQuantities which handle the indices of charge carriers on different regions.

chargeCarrierList::Union{Vector{VoronoiFVM.AbstractQuantity}, Vector{Int64}}

This list stores all charge carriers. Here, we can have a vector holding all abstract quantities or a vector holding an integer array depending on the interface model and the the regularity of unknowns.

indexPsi::Union{Int64, VoronoiFVM.AbstractQuantity}

This variable stores the index of the electric potential. Based on the user choice we have with this new type the opportunity to simulate with Quantities or integer indices.

params::ChargeTransportParams

A struct holding all region dependent parameter information. For more information see struct ChargeTransportParams.

paramsnodal::ChargeTransportParamsNodal

A struct holding all space dependent parameter information. For more information see struct ChargeTransportParamsNodal.

ChargeTransportInSolids.ChargeTransportData — Method

ChargeTransportData(grid, numberOfCarriers) -> ChargeTransportData

Simplified constructor for ChargeTransportData which only takes the grid and the numberOfCarriers as argument. Here, all necessary information including the physical parameters, but also some numerical information are located.

ChargeTransportInSolids.ChargeTransportIonicChargeCarriers — Type

mutable struct ChargeTransportIonicChargeCarriers

A struct holding all information necessary on the ionic charge carriers. With help of this constructor we can read out the indices the user chooses for ionic charge carrier quasi Fermi potentials and the respective regions in which they are defined. Note that it is possible to use ions as well as ion vacancies.

ionic_carriers::Vector{Int64}

Array with the indices of ionic charge carriers.

regions::Vector{Int64}

Corresponding regions where ionic charge carriers are assumed to be present.

ChargeTransportInSolids.ChargeTransportParams — Type

mutable struct ChargeTransportParams

A struct holding the physical region dependent parameters for a drift-diffusion simulation of a semiconductor device.

numberOfNodes::Int64

number of nodes used for the disretization of the domain $\mathbf{\Omega}$.

numberOfRegions::Int64

number of subregions $\mathbf{\Omega}_k$ within the domain $\mathbf{\Omega}$.

numberOfBoundaryRegions::Int64

number of boundary regions $(\partial \mathbf{\Omega})_k$ such that $\partial \mathbf{\Omega} = \cup_k (\partial \mathbf{\Omega})_k$.

numberOfCarriers::Int64

number of moving charge carriers.

numberOfInterfaceCarriers::Int64

number of present interface carriers which solely live on the boundary.

temperature::Float64

A given constant temperature.

UT::Float64

The thermal voltage, which reads $U_T = k_B T / q$.

γ::Float64

The parameter of the Blakemore statistics.

r0::Float64

Prefactor of electro-chemical reaction of internal boundary conditions.

SchottkyBarrier::Vector{Float64}

An array for the given Schottky barriers at present Schotkky contacts.

chargeNumbers::Vector{Float64}

An array with the corresponding charge numbers $z_\alpha$ for all carriers $\alpha$.

bBandEdgeEnergy::Matrix{Float64}

An array with the corresponding boundary band-edge energy values $E_\alpha$ in each region for each carrier $\alpha$.

bDensityOfStates::Matrix{Float64}

An array with the corresponding boundary effective density of states values $N_\alpha$ for each carrier $\alpha$.

bMobility::Matrix{Float64}

A 2D array with the corresponding boundary mobility values $\mu_\alpha$ in each region for each carrier $\alpha$.

bDoping::Matrix{Float64}

A 2D array with the corresponding boundary doping values for each carrier $\alpha$.

bVelocity::Matrix{Float64}

A 2D array with the corresponding boundary velocity values for each carrier $\alpha$, when assuming Schottky contacts.

recombinationSRHvelocity::Matrix{Float64}

A 2D array with the corresponding recombination surface boundary velocity values for electrons and holes.

bRecombinationSRHTrapDensity::Matrix{Float64}

A 2D array with the corresponding recombination surface boundary density values for electrons and holes.

doping::Matrix{Float64}

A 2D array with the corresponding doping values for each carrier $\alpha$ on each region.

densityOfStates::Matrix{Float64}

A 2D array with the corresponding effective density of states values $N_\alpha$ for each carrier $\alpha$ on each region.

bandEdgeEnergy::Matrix{Float64}

A 2D array with the corresponding band-edge energy values $E_\alpha$ for each carrier $\alpha$ on each region.

mobility::Matrix{Float64}

A 2D array with the corresponding mobility values $\mu_\alpha$ for each carrier $\alpha$ on each region.

recombinationSRHLifetime::Matrix{Float64}

A 2D array with the corresponding SRH lifetimes $\tau_n, \tau_p$ for electrons and holes.

recombinationSRHTrapDensity::Matrix{Float64}

A 2D array with the corresponding SRH trap densities $n_{\tau}, p_{\tau}$ for electrons and holes.

recombinationAuger::Matrix{Float64}

A 2D array with the corresponding Auger coefficients for electrons and holes.

dielectricConstant::Vector{Float64}

A region dependent dielectric constant.

generationIncidentPhotonFlux::Vector{Float64}

A region dependent array for the prefactor in the generation process which is the incident photon flux.

generationUniform::Vector{Float64}

A region dependent array for an uniform generation rate.

generationAbsorption::Vector{Float64}

A region dependent array for the absorption coefficient in the generation process.

recombinationRadiative::Vector{Float64}

A region dependent array for the radiative recombination rate.

ChargeTransportInSolids.ChargeTransportParams — Method

ChargeTransportParams(grid, numberOfCarriers) -> ChargeTransportParams

Simplified constructor for ChargeTransportParams which only takes the grid and the numberOfCarriers as argument.

ChargeTransportInSolids.ChargeTransportParamsNodal — Type

mutable struct ChargeTransportParamsNodal

A struct holding the physical nodal, i.e. space-dependent parameters for a drift-diffusion simulation of a semiconductor device.

dielectricConstant::Vector{Float64}

A node dependent dielectric constant.

doping::Vector{Float64}

A 1D array with the corresponding doping values on each node.

mobility::Matrix{Float64}

A 2D array with the corresponding mobility values $\mu_\alpha$ for each carrier $\alpha$ on each node.

densityOfStates::Matrix{Float64}

A 2D array with the corresponding effective density of states values $N_\alpha$ for each carrier $\alpha$ on each node.

bandEdgeEnergy::Matrix{Float64}

A 2D array with the corresponding band-edge energy values $E_\alpha$ for each carrier $\alpha$ on each node.

ChargeTransportInSolids.ChargeTransportParamsNodal — Method

ChargeTransportParamsNodal(grid, numberOfCarriers) -> ChargeTransportParamsNodal

Simplified constructor for ChargeTransportParamsNodal which only takes the grid and the numberOfCarriers as argument.

ChargeTransportInSolids.ChargeTransportSystem — Type

mutable struct ChargeTransportSystem

A struct holding all information necessary for a drift-diffusion type system.

data::ChargeTransportData

A struct holding all data information, see ChargeTransportData

fvmsys::VoronoiFVM.AbstractSystem

A struct holding system information for the finite volume system.

ChargeTransportInSolids.ChargeTransportSystem — Method

ChargeTransportSystem(grid, data; unknown_storage)

System constructor which builds all necessary information needed based on the input parameters with special regard to additional interface models. This is the main struct in which all information on the input data, but also on the solving system, with which the calculations are performed, are stored.

ChargeTransportInSolids.ChargeTransportTraps — Type

mutable struct ChargeTransportTraps

A struct holding all information necessary for enabling traps in the SRH recombination. With help of this constructor we can read out the index the user chooses for trap quasi Fermi potentials and the respective regions in which they are defined.

traps::Int64

Array with the index of traps.

regions::Vector{Int64}

Corresponding regions where traps are assumed to be present.

ChargeTransportInSolids.abstract_model_SRH — Type

abstract type abstract_model_SRH

Abstract type for SRH bulk recombination model

!!! compat  
This one will be removed in future versions.

ChargeTransportInSolids.boundary_model — Type

abstract type boundary_model

Abstract type for boundary model. Subtypes are ohmiccontact, schottkycontact and interface_model.

ChargeTransportInSolids.calculation_type — Type

abstract type calculation_type

Abstract type calculation_type which distinguishes between equilibrium and out of equilibrium calculations.

ChargeTransportInSolids.diffusion_enhanced — Type

abstract type diffusion_enhanced <: ChargeTransportInSolids.flux_approximation

Abstract type for diffusion enhanced flux discretization.

ChargeTransportInSolids.excess_chemical_potential — Type

abstract type excess_chemical_potential <: ChargeTransportInSolids.flux_approximation

Abstract type for excess chemical potential flux discretization.

ChargeTransportInSolids.excess_chemical_potential_graded — Type

abstract type excess_chemical_potential_graded <: ChargeTransportInSolids.flux_approximation

Abstract type for excess chemical potential flux discretization for graded effective density of states and/or graded band-edge energies. This means, use this flux when at least one of these quantities is assumed to be space-dependent.

ChargeTransportInSolids.flux_approximation — Type

abstract type flux_approximation

Abstract type for flux discretization model.

ChargeTransportInSolids.generalized_sg — Type

abstract type generalized_sg <: ChargeTransportInSolids.flux_approximation

Abstract type for generalized Scharfetter-Gummel flux discretization. This flux approximation results in an implicit equation which needs to be solved and is exact for all Blakemore type statistics functions with abritary γ.

ChargeTransportInSolids.generation_beer_lambert — Type

abstract type generation_beer_lambert <: generation_model

Abstract type for Beer-Lambert generation.

ChargeTransportInSolids.generation_model — Type

abstract type generation_model

Abstract type for generation model.

ChargeTransportInSolids.generation_none — Type

abstract type generation_none <: generation_model

Abstract type for no generation model.

ChargeTransportInSolids.generation_uniform — Type

abstract type generation_uniform <: generation_model

Abstract type for uniform generation.

ChargeTransportInSolids.inEquilibrium — Type

abstract type inEquilibrium <: ChargeTransportInSolids.calculation_type

Abstract type for equilibrium calculations.

ChargeTransportInSolids.interface_model — Type

abstract type interface_model <: boundary_model

Abstract type for interface model which is part of boundary model. Subtypes are given below.

ChargeTransportInSolids.interface_model_discont_qF — Type

abstract type interface_model_discont_qF <: interface_model

Abstract type for an interface model where discontinuous quasi Fermi potentials are needed.

ChargeTransportInSolids.interface_model_ion_charge — Type

abstract type interface_model_ion_charge <: interface_model

Abstract type for present ion charges at interfaces.

ChargeTransportInSolids.interface_model_none — Type

abstract type interface_model_none <: interface_model

Abstract type for no interface model.

ChargeTransportInSolids.interface_model_surface_recombination — Type

abstract type interface_model_surface_recombination <: interface_model

Abstract type for surface recombination mechanisms.

ChargeTransportInSolids.linearScanProtocol — Type

abstract type linearScanProtocol <: scan_protocol_type

Abstract type for linear scan protocol.

ChargeTransportInSolids.model_SRH — Type

abstract type model_SRH <: abstract_model_SRH

model_SRH as parent of several different subtypes.

ChargeTransportInSolids.model_SRH_2species_present_trap_dens — Type

abstract type model_SRH_2species_present_trap_dens <: abstract_model_SRH

This Datatype will be deleted soon.

ChargeTransportInSolids.model_SRH_with_traps — Type

abstract type model_SRH_with_traps <: model_SRH

modelSRHwith_traps as parent of several different subtypes.

ChargeTransportInSolids.model_SRH_without_traps — Type

abstract type model_SRH_without_traps <: model_SRH

modelSRHwithout_traps as parent of several different subtypes.

ChargeTransportInSolids.model_stationary — Type

abstract type model_stationary <: model_type

Abstract type for stationary simulations.

ChargeTransportInSolids.model_transient — Type

abstract type model_transient <: model_type

Abstract type for transient simulations.

ChargeTransportInSolids.model_type — Type

abstract type model_type

Abstract type for model type which indicates, if we consider stationary or transient problem.

ChargeTransportInSolids.ohmic_contact — Type

Abstract type for ohmic contacts as boundary model.

ChargeTransportInSolids.outOfEquilibrium — Type

abstract type outOfEquilibrium <: ChargeTransportInSolids.calculation_type

Abstract type for out of equilibrium calculations.

ChargeTransportInSolids.scan_protocol_type — Type

abstract type scan_protocol_type

Abstract type for scan protocol type

ChargeTransportInSolids.scharfetter_gummel — Type

abstract type scharfetter_gummel <: ChargeTransportInSolids.flux_approximation

Abstract type for Scharfetter-Gummel flux discretization. Choose this one, when the Boltzmann statistics function is chosen as statistics.

ChargeTransportInSolids.scharfetter_gummel_graded — Type

abstract type scharfetter_gummel_graded <: ChargeTransportInSolids.flux_approximation

Abstract type for Scharfetter-Gummel flux discretization for graded effective density of states and/or graded band-edge energies. This means, use this flux when at least one of these quantities is assumed to be space-dependent.

ChargeTransportInSolids.schottky_contact — Type

Abstract type for schottky contacts as boundary model.

Methods

ChargeTransportInSolids.Blakemore — Method

Blakemore(x::Real, γ::Real) -> Any

The Blakemore approximation $1/(\exp(-x) + γ)$ with variable real scalar $γ$, see [Blakemore1952, "The parameters of partially degenerate semiconductors"].

ChargeTransportInSolids.Blakemore — Method

Blakemore(x::Real) -> Any

The Blakemore approximation $1/(\exp(-x) + γ)$ with $γ = 0.27$, see [Blakemore1952, "The parameters of partially degenerate semiconductors"].

ChargeTransportInSolids.Boltzmann — Method

Boltzmann(x::Real) -> Any

The Boltzmann statistics function $\exp(x)$.

ChargeTransportInSolids.FermiDiracMinusOne — Method

FermiDiracMinusOne(x::Real) -> Any

The Fermi-Dirac integral of order $-1$ which reads $1/(\exp(-x) + 1)$, see [Blakemore1982: "Approximations for Fermi-Dirac integrals, especially the function $F_{1/2} (η)$ used to describe electron density in a semiconductor"].

ChargeTransportInSolids.FermiDiracOneHalfBednarczyk — Method

FermiDiracOneHalfBednarczyk(x::Real) -> Any

The incomplete Fermi-Dirac integral of order 1/2, implemented according to [Bednarczyk1978, "The Approximation of the Fermi-Dirac integral $F_{1/2}()$"].

ChargeTransportInSolids.FermiDiracOneHalfTeSCA — Method

FermiDiracOneHalfTeSCA(x::Real) -> Any

The incomplete Fermi-Dirac integral of order 1/2, implemented according to the software package TeSCA, see https://wias-berlin.de/software/index.jsp?lang=1&id=TeSCA.

ChargeTransportInSolids.addChargeCarriers! — Method

addChargeCarriers!(f, u, node, data, ipsi, iphin, iphip, n, p) -> Any

Function which builds right-hand side of charge carriers.

ChargeTransportInSolids.addElectricPotential! — Method

addElectricPotential!(f, u, node, data, ipsi) -> Any

Function which builds right-hand side of Poisson equation, i.e. which builds the space charge density for outOfEquilibrium calculations.

ChargeTransportInSolids.addTrapDensity! — Method

addTrapDensity!(f, u, node, data, ipsi, iphin, iphip, n, p) -> Any

Function which adds additional trap density to right-hand side of Poisson equation without modeling traps as own charge carrier. Note that this one may be deleted in future version.

ChargeTransportInSolids.bflux! — Method

bflux!(f, u, bedge, data) -> Any

Master bflux! function. This is the function which enters VoronoiFVM and hands over for each boundary the flux within the boundary.

ChargeTransportInSolids.bflux! — Method

bflux!(f, u, bedge, data, _::Type{excess_chemical_potential}) -> Any

The excess chemical potential flux discretization scheme for inner boundaries.

ChargeTransportInSolids.breaction! — Method

breaction!(f, u, bnode, data) -> Any

Master breaction! function. This is the function which enters VoronoiFVM and hands over for each boundary the chosen boundary model.

ChargeTransportInSolids.breaction! — Method

breaction!(f, u, bnode, data, _::Type{interface_model_none})

This breaction! function is chosen when no interface model is chosen.

ChargeTransportInSolids.breaction! — Method

breaction!(f, u, bnode, data, _::Type{ohmic_contact}) -> Any

Creates ohmic boundary conditions via a penalty approach with penalty parameter $\delta$. For example, the right-hand side for the electrostatic potential $\psi$ is implemented as

$f[\psi] = -q/\delta ( (p - N_a) - (n - N_d) )$,

assuming a bipolar semiconductor. In general, we have for some given charge number $z_\alpha$

$f[\psi] = -q/\delta \sum_\alpha{ z_\alpha (n_\alpha - C_\alpha) },$

where $C_\alpha$ corresponds to some doping w.r.t. the species $\alpha$.

The boundary conditions for the charge carriers are set in the main file. Hence,

$f[n_\alpha] = 0$`

for all charge carriers $\alpha$.

ChargeTransportInSolids.breaction! — Method

breaction!(f, u, bnode, data, _::Type{schottky_contact}) -> Any

[Note that this way of implementation is not well tested yet.

Creates Schottky boundary conditions in a first attempt. For the electrostatic potential we assume

$\psi = \psi_S + U,$

where $\psi_S$ corresponds to a given value and $U$ to the applied voltage. For now, the quantitity $\psi_S$ needs to be specified in the main file. For the charge carriers we assume the following

$f[n_\alpha] = z_\alpha q v_\alpha (n_\alpha - n_{\alpha, 0})$,

where $v_{\alpha}$ can be treated as a surface recombination mechanism and is given. The parameter $n_{\alpha, 0}$ is a given value, calculated by the statistical relation, when assuming no electrical field and a quasi Fermi level equal to the metal work function $\phi$, i.e.

$n_{\alpha, 0}= z_\alpha/ U_T (E_\alpha - \phi) / q.$

ChargeTransportInSolids.bstorage! — Method

bstorage!(f, u, bnode, data) -> Any

Master bstorage! function. This is the function which enters VoronoiFVM and hands over for each boundary the time-dependent part of the chosen boundary model.

ChargeTransportInSolids.bstorage! — Method

bstorage!(f, u, bnode, data, _::Type{interface_model_discont_qF})

No bstorage! is used, when assuming discontinuous qF.

ChargeTransportInSolids.bstorage! — Method

bstorage!(f, u, bnode, data, _::Type{ohmic_contact})

No bstorage! is used, if an ohmic and schottky contact model is chosen.

ChargeTransportInSolids.build_system — Method

build_system(grid, data, unknown_storage, _)

The core of the new system constructor. Here, the system for discontinuous quantities is build.

ChargeTransportInSolids.build_system — Method

build_system(grid, data, unknown_storage, _)

The core of the system constructor. Here, the system for no additional interface model is build.

ChargeTransportInSolids.chargeDensity — Method

chargeDensity(psi0, phi, UT, EVector, chargeNumbers, dopingVector, dosVector, FVector) -> Any

Compute the charge density, i.e. the right-hand side of Poisson's equation.

ChargeTransportInSolids.chargeDensity — Method

chargeDensity(ctsys, sol) -> Any

Compute the charge density for each region separately.

ChargeTransportInSolids.compute_densities! — Method

compute_densities!(u, data, inode, region, icc, ipsi, in_region::Bool) -> Any

For given potentials, compute corresponding densities. This function is needed for the method, plotting the densities.

ChargeTransportInSolids.compute_densities! — Method

compute_densities!(grid, data, sol) -> Matrix{Real}

For given potentials in vector form, compute corresponding vectorized densities. [Caution: this was not tested for multidimensions.]

ChargeTransportInSolids.compute_energies! — Method

compute_energies!(grid, data, sol)

For given solution in vector form, compute corresponding vectorized band-edge energies and Fermi level. [Caution: this was not tested for multidimensions.]

ChargeTransportInSolids.degenerateLimit — Method

degenerateLimit(x) -> Any

Degenerate limit of incomplete Fermi-Dirac integral of order 1/2.

ChargeTransportInSolids.electroNeutralSolution! — Method

electroNeutralSolution!(grid, data; Newton) -> Any

Compute the electro-neutral solution for the Boltzmann approximation. It is obtained by setting the left-hand side in the Poisson equation equal to zero and solving for $\psi$. The charge carriers may obey different statitics functions. Currently, this one is not well tested for the case of charge carriers beyond electrons and holes.

ChargeTransportInSolids.electrochemicalReaction — Method

electrochemicalReaction(data, u, iphia, ipsi, iphiaJunction, ipsiJunction, β, κ, DOS, E) -> Any

Electrochemical reaction between interface and bulk ionic species. This function enters in the internal boundary reaction in case of an ion charge interface model.

ChargeTransportInSolids.enable_ionic_carriers — Method

Corresponding constructor for the present ionic charge carriers and the respective regions.

ChargeTransportInSolids.enable_traps! — Method

Corresponding constructor for the present trap density and the respective regions. Note that, this one will may be deleted in future versions.

ChargeTransportInSolids.enable_traps! — Method

Corresponding constructor for the present trap density and the respective regions.

ChargeTransportInSolids.etaFunction — Method

etaFunction(psi, phi, UT, E, z) -> Any

The argument of the distribution function for given $\varphi_\alpha$ and $\psi$

$z_\alpha / U_T ( (\varphi_\alpha - \psi) + E_\alpha / q ).$

The parameters $E_\alpha$ and $z_\alpha$ are given as vectors. This function may be used to compute the charge density, i.e. the right-hand side of the Poisson equation.

ChargeTransportInSolids.etaFunction — Method

etaFunction(u, data, node, region, icc, ipsi, in_region::Bool) -> Any

The argument of the distribution function

$z_\alpha / U_T ( (\varphi_\alpha - \psi) + E_\alpha / q )$

for floats.

ChargeTransportInSolids.etaFunction — Method

etaFunction(u, bnode::VoronoiFVM.BNode, data, icc, ipsi) -> Any

The argument of the distribution function

$z_\alpha / U_T ( (\varphi_\alpha - \psi) + E_\alpha / q )$

for boundary nodes.

ChargeTransportInSolids.etaFunction — Method

etaFunction(u, node::VoronoiFVM.Node, data, icc, ipsi) -> Any

The argument of the distribution function

$z_\alpha / U_T ( (\varphi_\alpha - \psi) + E_\alpha / q )$

for interior nodes.

ChargeTransportInSolids.flux! — Method

flux!(f, u, edge, data, _::Type{diffusion_enhanced}) -> Any

The diffusion enhanced scheme by Bessemoulin-Chatard. Currently, the Pietra-Jüngel scheme is used for the regularization of the removable singularity. This also works for space-dependent band-edge energy, but not for space-dependent effective DOS.

ChargeTransportInSolids.flux! — Method

flux!(f, u, edge, data, _::Type{excess_chemical_potential_graded}) -> Any

The excess chemical potential flux scheme for possible space-dependent DOS and band-edge energies. For these parameters the discretization scheme is modified. [insert continuous flux etc ...]

ChargeTransportInSolids.flux! — Method

flux!(f, u, edge, data, _::Type{excess_chemical_potential}) -> Any

The excess chemical potential flux discretization scheme. This also works for space-dependent band-edge energy, but not for space-dependent effective DOS.

ChargeTransportInSolids.flux! — Method

flux!(f, u, edge, data, _::Type{generalized_sg}) -> Any

The Koprucki-Gärtner scheme. This scheme is calculated by solving a fixed point equation which arise when considering the generalized Scharfetter-Gummel scheme in case of Blakemore statistics. Hence, it should be exclusively worked with, when considering the Blakemore distribution. This also works for space-dependent band-edge energy, but not for space-dependent effective DOS.

ChargeTransportInSolids.flux! — Method

flux!(f, u, edge, data, _::Type{scharfetter_gummel_graded}) -> Any

The classical Scharfetter-Gummel flux scheme for possible space-dependent DOS and band-edge energies. For these parameters the discretization scheme is modified. [insert continuous flux etc ...]

ChargeTransportInSolids.flux! — Method

flux!(f, u, edge, data, _::Type{scharfetter_gummel}) -> Any

The classical Scharfetter-Gummel flux scheme. This also works for space-dependent band-edge energy, but not for space-dependent effective DOS.

ChargeTransportInSolids.generation — Method

generation(data, ireg, node, _::Type{generation_beer_lambert}) -> Any

The generation rate $G$, which occurs in the right-hand side of the continuity equations obeying the Beer-Lambert law.

ChargeTransportInSolids.generation — Method

generation(data, ireg, node, _::Type{generation_uniform}) -> Any

The generation rate $G$, which occurs in the right-hand side of the continuity equations with a uniform generation rate.

ChargeTransportInSolids.get_BEE! — Method

get_BEE!(icc, bnode::VoronoiFVM.BNode, data) -> Any

Defining locally the band-edge energy for boundary nodes.

ChargeTransportInSolids.get_BEE! — Method

get_BEE!(icc, edge::VoronoiFVM.Edge, data) -> Any

Defining locally the band-edge energy for edges.

ChargeTransportInSolids.get_BEE! — Method

get_BEE!(icc, node::VoronoiFVM.Node, data) -> Any

Defining locally the band-edge energy for interior nodes.

ChargeTransportInSolids.get_DOS! — Method

get_DOS!(icc, bnode::VoronoiFVM.BNode, data) -> Any

Defining locally the effective DOS for boundary nodes.

ChargeTransportInSolids.get_DOS! — Method

get_DOS!(icc, edge::VoronoiFVM.Edge, data) -> Tuple{Any, Any}

Defining locally the effective DOS for edges.

ChargeTransportInSolids.get_DOS! — Method

get_DOS!(icc, node::VoronoiFVM.Node, data) -> Any

Defining locally the effective DOS for interior nodes.

ChargeTransportInSolids.get_current_val — Method

Calculates current for time dependent problem. But caution, still need some small modification!

ChargeTransportInSolids.get_current_val — Method

Calculates current for stationary problem. But caution, still need some small modification!

ChargeTransportInSolids.inner_interface_model — Method

inner_interface_model(data)

Method which determines with input parameters which inner interface model was chosen by user.

ChargeTransportInSolids.inner_interface_model — Method

inner_interface_model(ctsys)

Method which determines with input parameters which inner interface model was chosen by user.

ChargeTransportInSolids.kernelSRH — Method

kernelSRH(data, ireg, iphin, iphip, n, p, _::Type{ChargeTransportInSolids.model_SRH_stationary}) -> Any

SRH kernel for case of using stationary formula, i.e. case where no present traps are assumed.

ChargeTransportInSolids.kernelSRH — Method

kernelSRH(data, ireg, iphin, iphip, n, p, _::Type{model_SRH_off}) -> Float64

SRH kernel for case of non-existing rate.

ChargeTransportInSolids.plotDiffusionEnhancements — Method

plotDiffusionEnhancements()

Plot diffusion enhancements.

ChargeTransportInSolids.plotDistributions — Method

plotDistributions(; Plotter)

Plot different distribution integrals.

ChargeTransportInSolids.plot_IV — Method

plot_IV(Plotter, biasValues, IV, Δu; plotGridpoints) -> Any

Method for showing the total current in dependence of the applied voltage. One input parameter is the boolean plotGridpoints which makes it possible to plot markers, which indicate where the nodes are located.

ChargeTransportInSolids.plot_densities — Method

plot_densities(Plotter, grid, data, sol, title; plotGridpoints) -> Any

Plotting routine, where the charge carrier densities are depicted in dependence of space. The case of heterojunctions is tested, but yet multidimensional calculations are missing. Currently, for a matching legend, we need the following order:

index 1: electrons as charge carrier with the corresponding density $n$,

index 2: holes as charge carrier with the corresponding density $p$,

index 3: anion vacancies as charge carrier with the corresponding density $a$,

index 4: cation vacancies as charge carrier with the corresponding density $c$.

One input parameter is the boolean plotGridpoints which makes it possible to plot markers, which indicate where the nodes are located.

ChargeTransportInSolids.plot_doping — Method

plot_doping(Plotter, g::ExtendableGrids.ExtendableGrid, data::ChargeTransportData) -> Any

Possibility to plot the considered doping. This is especially useful for making sure that the interior and the boundary doping agree.

ChargeTransportInSolids.plot_doping — Method

Plot doping for nodal dependent doping

ChargeTransportInSolids.plot_electroNeutralSolutionBoltzmann — Method

plot_electroNeutralSolutionBoltzmann(Plotter, grid, psi0; plotGridpoints) -> Any

Plotting routine for depicting the electroneutral potential. One input parameter is the boolean plotGridpoints which makes it possible to plot markers, which indicate where the nodes are located.

ChargeTransportInSolids.plot_energies — Method

plot_energies(Plotter, grid, data, sol, title; plotGridpoints) -> Any

With this method it is possible to plot the energies

$E_\alpha - q \psi \quad \text{w.r.t. space.}$

The case of heterojunctions is tested, but yet multidimensional calculations are missing. Currently, for a matching legend, we need the following order:

index 1: electrons as charge carrier with the corresponding density $n$,

index 2: holes as charge carrier with the corresponding density $p$,

index 3: anion vacancies as charge carrier with the corresponding density $a$,

index 4: cation vacancies as charge carrier with the corresponding density $c$.

One input parameter is the boolean plotGridpoints which makes it possible to plot markers, which indicate where the nodes are located.

ChargeTransportInSolids.plot_energies — Method

plot_energies(Plotter, grid, data)

With this method it is possible to depict the band-edge energies $E_\alpha$. This can be useful for debugging when dealing with heterojunctions.

ChargeTransportInSolids.plot_solution — Method

plot_solution(Plotter, grid, data, solution, title; plotGridpoints) -> Any

Method for plotting the solution vectors: the electrostatic potential $\psi$ as well as the charge carriers. The case of heterojunctions is tested, but yet multidimensional calculations are missing. Currently, for a matching legend, we need the following order:

index 1: electrons as charge carrier with the corresponding density $n$,

index 2: holes as charge carrier with the corresponding density $p$,

index 3: anion vacancies as charge carrier with the corresponding density $a$,

index 4: cation vacancies as charge carrier with the corresponding density $c$. One input parameter is the boolean plotGridpoints which makes it possible to plot markers, which indicate where the nodes are located.

ChargeTransportInSolids.printJacobi — Method

printJacobi(node, sys)

First try of debugger. Print the Jacobi matrix for a given node, i.e. the number of the node in the grid and not the excact coordinate. This is only done for the one dimensional case so far. [insert some information about VoronoiFVM]

ChargeTransportInSolids.reaction! — Method

reaction!(f, u, node, data) -> Any

Master reaction! function. This is the function which enters VoronoiFVM and hands over reaction terms for concrete calculation type and bulk recombination model.

ChargeTransportInSolids.reaction! — Method

reaction!(f, u, node, data, _::Type{inEquilibrium}) -> Any

Reaction in case of equilibrium, i.e. no generation and recombination is considered.

ChargeTransportInSolids.reaction! — Method

reaction!(f, u, node, data, _::Type{outOfEquilibrium}) -> Any

Sets up the right-hand sides. Assuming a bipolar semiconductor the right-hand side for the electrostatic potential becomes

$f[ψ] = - q ((p - N_a) - (n - N_d) ) = - q \sum n_\alpha (n_\alpha - C_\alpha)$

for some doping $C_\alpha$ w.r.t. to the species $\alpha$. The right-hand sides for the charge carriers read as

$f[n_\alpha] = - z_\alpha q (G - R)$

for all charge carriers $n_\alpha$. The recombination includes radiative, Auger and Shockley-Read-Hall recombination. For latter recombination process the stationary simplification is implemented.

The recombination is only implemented for electron and holes and assumes that the electron index is 1 and the hole index is 2.

ChargeTransportInSolids.set_bulk_recombination — Method

set_bulk_recombination(; iphin, iphip, bulk_recomb_Auger, bulk_recomb_radiative, bulk_recomb_SRH)

Corresponding constructor for the bulk recombination model.

ChargeTransportInSolids.set_time_mesh — Method

Gives the user a linear time mesh, this mesh is used for a linear I-V scan protocol.

ChargeTransportInSolids.storage! — Method

storage!(f, u, node, data) -> Any

Master storage! function. This is the function which enters VoronoiFVM and hands over a storage term, if we consider transient problem.

ChargeTransportInSolids.storage! — Method

storage!(f, u, node, data, _::Type{model_transient}) -> Any

The storage term for time-dependent problems. Currently, for the time-dependent current densities the implicit Euler scheme is used. Hence, we have

$f[n_\alpha] = z_\alpha q ∂_t n_\alpha$

and for the electrostatic potential $f[ψ] = 0$.

ChargeTransportInSolids.trap_density! — Method

trap_density!(icc, ireg, data, Et)

Compute trap densities for a given trap energy. [Currently, only done for the Boltzmann statistics and for region dependent parameters.]